# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/mission/{id}/targets": {
            "post": {
                "description": "Adds multiple targets to a specific mission.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "Add targets to a mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Targets Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sca_internal_service_target.AddTargetsToMissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Targets added to mission successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/target/{id}": {
            "delete": {
                "description": "Deletes a specific target.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "Delete a target",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/target/{id}/completion": {
            "put": {
                "description": "Updates the completion status of a target.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "Update target completion",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Completion Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sca_internal_service_target.UpdateTargetCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target completion updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/target/{id}/notes": {
            "put": {
                "description": "Updates the notes of a target.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "Update target notes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Notes Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sca_internal_service_target.UpdateTargetNotesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target notes updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "sca_internal_service_target.AddTargetsToMissionRequest": {
            "type": "object",
            "required": [
                "mission_id",
                "targets"
            ],
            "properties": {
                "mission_id": {
                    "type": "integer"
                },
                "targets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sca_internal_service_target.Target"
                    }
                }
            }
        },
        "sca_internal_service_target.Target": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "mission_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "sca_internal_service_target.UpdateTargetCompletionRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_completed": {
                    "type": "boolean"
                }
            }
        },
        "sca_internal_service_target.UpdateTargetNotesRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                }
            }
        }
    }
}